requirements.txt
discord.py
humanize
jishaku
httpx
ujson
regex
yarl